LF-14347: armv8/ls216xardb: sync RCWSR12 status in DCSR with CCSR
DCSR is the shadow register of CCSR, but the values of these registers
are not updated simultaneously. When the RCWSR12 register in DCSR is
operated through the bit mask to change the I2C5 PINMUX, wrong value will
overwrite the USDHC0/CAN0/CAN1 PINMUX bit in CCSR RCWSR12 register. Sync
RCWSR12 register value in DCSR with CCSR to avoid the defect.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>